fix: use real(dp) for tolerance in results_equals
The confidence_score field was updated to real(dp) but the local
tolerance parameter in results_equals was still using plain real.
This ensures consistent precision for the comparison.

+++ src/semantic/analyzers/base_analyzer.f90
@@ -57,7 +57,7 @@ contains
     function results_equals(this, other) result(equal)
         class(analysis_results_t), intent(in) :: this, other
         logical :: equal
-        real, parameter :: tolerance = 1.0e-6
+        real(dp), parameter :: tolerance = 1.0d-6
 
         equal = this%converged .eqv. other%converged .and. &
                 this%iteration_count == other%iteration_count .and. &
